Successful sound play and lock test in testLevel
I figured out how to get a stem to kick in on collision and to keep
playing thereafter. Dr 0. suggested that in order to avoid timing
issues, we should have all sounds playing on mute, and the collision
will then unmute it. I have no idea how to make it mute and unmute, and
the tutorials have failed me for the first time.